MAINT-1416 Move moduleId to abstract SnomedComponent class.
