+++ src/main/java/org/snomed/snowstorm/core/data/domain/Concept.java
@@ -28,7 +28,6 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 
 	public interface Fields extends SnomedComponent.Fields {
 		String CONCEPT_ID = "conceptId";
-		String MODULE_ID = "moduleId";
 		String DEFINITION_STATUS_ID = "definitionStatusId";
 	}
 
@@ -50,12 +49,6 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 	@JsonIgnore
 	// Populated when requesting an update
 	private Map<String, Set<String>> associationTargetStrings;
-
-	@JsonView(value = View.Component.class)
-	@Field(type = FieldType.keyword)
-	@NotNull
-	@Size(min = 5, max = 18)
-	private String moduleId;
 
 	@Field(type = FieldType.keyword)
 	@NotNull
@@ -83,7 +76,7 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 
 	public Concept() {
 		active = true;
-		moduleId = Concepts.CORE_MODULE;
+		setModuleId(Concepts.CORE_MODULE);
 		definitionStatusId = Concepts.PRIMITIVE;
 		descriptions = new HashSet<>();
 		relationships = new HashSet<>();
@@ -100,7 +93,7 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 	public Concept(String conceptId, String moduleId) {
 		this();
 		this.conceptId = conceptId;
-		this.moduleId = moduleId;
+		setModuleId(moduleId);
 	}
 
 	public Concept(String conceptId, Integer effectiveTime, boolean active, String moduleId, String definitionStatusId) {
@@ -108,7 +101,7 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 		this.conceptId = conceptId;
 		setEffectiveTimeI(effectiveTime);
 		this.active = active;
-		this.moduleId = moduleId;
+		setModuleId(moduleId);
 		this.definitionStatusId = definitionStatusId;
 	}
 
@@ -121,7 +114,7 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 	public boolean isComponentChanged(Concept that) {
 		return that == null
 				|| active != that.active
-				|| !moduleId.equals(that.moduleId)
+				|| !getModuleId().equals(that.getModuleId())
 				|| !definitionStatusId.equals(that.definitionStatusId);
 	}
 
@@ -131,7 +124,7 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 
 	@Override
 	protected Object[] getReleaseHashObjects() {
-		return new Object[]{active, moduleId, definitionStatusId};
+		return new Object[]{active, getModuleId(), definitionStatusId};
 	}
 
 	@JsonView(value = View.Component.class)
@@ -234,7 +227,7 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 	}
 
 	public Concept addAxiom(Relationship... axiomFragments) {
-		Axiom axiom = new Axiom(moduleId, true, Concepts.PRIMITIVE, Sets.newHashSet(axiomFragments));
+		Axiom axiom = new Axiom(getModuleId(), true, Concepts.PRIMITIVE, Sets.newHashSet(axiomFragments));
 		axiom.getRelationships().forEach(r -> r.setSourceId(this.conceptId));
 		classAxioms.add(axiom);
 		return this;
@@ -247,7 +240,7 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 	}
 
 	public Concept addGeneralConceptInclusionAxiom(Relationship... axiomFragments) {
-		Axiom axiom = new Axiom(moduleId, true, Concepts.PRIMITIVE, Sets.newHashSet(axiomFragments));
+		Axiom axiom = new Axiom(getModuleId(), true, Concepts.PRIMITIVE, Sets.newHashSet(axiomFragments));
 		axiom.getRelationships().forEach(r -> r.setSourceId(this.conceptId));
 		generalConceptInclusionAxioms.add(axiom);
 		return this;
@@ -333,16 +326,6 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 	}
 
 	@Override
-	public String getModuleId() {
-		return moduleId;
-	}
-
-	public Concept setModuleId(String moduleId) {
-		this.moduleId = moduleId;
-		return this;
-	}
-
-	@Override
 	public String getDefinitionStatusId() {
 		return definitionStatusId;
 	}
@@ -413,13 +396,13 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 			return Objects.equals(conceptId, concept.conceptId);
 		}
 
-		return Objects.equals(moduleId, concept.moduleId) &&
+		return Objects.equals(getModuleId(), concept.getModuleId()) &&
 				Objects.equals(definitionStatusId, concept.definitionStatusId);
 	}
 
 	@Override
 	public int hashCode() {
-		return Objects.hash(conceptId, moduleId, definitionStatusId);
+		return Objects.hash(conceptId, getModuleId(), definitionStatusId);
 	}
 
 	@Override
@@ -428,7 +411,7 @@ public class Concept extends SnomedComponent<Concept> implements ConceptView, Sn
 				"conceptId='" + conceptId + '\'' +
 				", effectiveTime='" + getEffectiveTimeI() + '\'' +
 				", active=" + active +
-				", moduleId='" + moduleId + '\'' +
+				", moduleId='" + getModuleId() + '\'' +
 				", definitionStatusId='" + definitionStatusId + '\'' +
 				", internalId='" + getInternalId() + '\'' +
 				", start='" + getStartDebugFormat() + '\'' +
